Refactor stack trace assertions
