+++ test/int/wizards/breakpoints/implementation/breakpointsAssertions.ts
@@ -1,4 +1,3 @@
-import * as assert from 'assert';
 import * as path from 'path';
 import { expect, use } from 'chai';
 import * as chaiString from 'chai-string';
@@ -8,20 +7,16 @@ import { BreakpointWizard } from '../breakpointWizard';
 import { InternalFileBreakpointsWizard, CurrentBreakpointsMapping } from './internalFileBreakpointsWizard';
 import { BreakpointsWizard } from '../breakpointsWizard';
 import { waitUntilReadyWithTimeout } from '../../../utils/waitUntilReadyWithTimeout';
-import { findLineNumber } from '../../../utils/findPositionOfTextInFile';
 import { IExpectedVariables, VariablesAssertions } from './variablesAssertions';
+import { ExpectedFrame, StackTraceObjectAssertions } from './stackTraceObjectAssertions';
+import { StackTraceStringAssertions } from './stackTraceStringAssertions';
 
 use(chaiString);
 
-export type IExpectedStackTrace = string;
-export interface IStackTraceVerifier {
-    format?: DebugProtocol.StackFrameFormat;
-    verify: (response: DebugProtocol.StackTraceResponse) => void;
-}
-
 export interface IVerifications {
     variables?: IExpectedVariables;
-    stackTrace?: IExpectedStackTrace | IStackTraceVerifier;
+    stackTrace?: string | ExpectedFrame[];
+    stackFrameFormat?: DebugProtocol.StackFrameFormat;
 }
 
 interface IObjectWithLocation {
@@ -46,31 +41,6 @@ export class BreakpointsAssertions {
         private readonly _internal: InternalFileBreakpointsWizard,
         public readonly currentBreakpointsMapping: CurrentBreakpointsMapping) { }
 
-    private getDefaultStackTraceVerifier(breakpoint: BreakpointWizard, expectedStackTrace: string): IStackTraceVerifier {
-        return {
-            format: this._defaultStackFrameFormat,
-            verify: (stackTraceResponse) => {
-                expect(stackTraceResponse.success, `Expected the response to the stack trace request to be succesful yet it failed: ${JSON.stringify(stackTraceResponse)}`).to.equal(true);
-
-                const stackTraceFrames = stackTraceResponse.body.stackFrames;
-                expect(stackTraceResponse.body.totalFrames, `The number of stackFrames was different than the value supplied on the totalFrames field`)
-                    .to.equal(stackTraceFrames.length);
-                stackTraceFrames.forEach(frame => {
-                    // Warning: We don't currently validate frame.source.path
-                    expect(frame.source).not.to.equal(undefined);
-                    const expectedSourceNameAndLine = ` [${frame.source!.name}] Line ${frame.line}`;
-                    expect(frame.name, 'Expected the formatted name to match the source name and line supplied as individual attributes').to.endsWith(expectedSourceNameAndLine);
-                });
-
-
-                const formattedExpectedStackTrace = expectedStackTrace.replace(/^\s+/gm, ''); // Remove the white space we put at the start of the lines to make the stack trace align with the code
-                this.applyIgnores(formattedExpectedStackTrace, stackTraceFrames);
-                const actualStackTrace = this.extractStackTrace(stackTraceFrames);
-                assert.equal(actualStackTrace, formattedExpectedStackTrace, `Expected the stack trace when hitting ${breakpoint} to be:\n${formattedExpectedStackTrace}\nyet it is:\n${actualStackTrace}`);
-            }
-        };
-    }
-
     public assertIsVerified(breakpoint: BreakpointWizard): void {
         // Convert to one based to match the VS Code potocol and what VS Code does if you try to open that file at that line number
 
@@ -94,19 +64,9 @@ export class BreakpointsAssertions {
     public async assertIsHitThenResume(breakpoint: BreakpointWizard, verifications: IVerifications): Promise<void> {
         await this._breakpointsWizard.waitUntilPaused(breakpoint);
 
-        let stackTraceVerifier: IStackTraceVerifier | undefined = undefined;
-        if (typeof verifications.stackTrace === 'string') {
-            stackTraceVerifier = this.getDefaultStackTraceVerifier(breakpoint, verifications.stackTrace);
-        } else if (typeof verifications.stackTrace === 'object') {
-            stackTraceVerifier = verifications.stackTrace;
-        }
+        const stackFrameFormat = verifications.stackFrameFormat || this._defaultStackFrameFormat;
 
-        let stackFrameFormat: DebugProtocol.StackFrameFormat | undefined = undefined;
-        if (stackTraceVerifier) {
-            stackFrameFormat = stackTraceVerifier.format;
-        }
-
-        const stackTraceResponse = await await this._internal.client.send('stackTrace', {
+        const stackTraceResponse = await this._internal.client.send('stackTrace', {
             threadId: THREAD_ID,
             format: stackFrameFormat
         });
@@ -115,8 +75,12 @@ export class BreakpointsAssertions {
         // Validate that the topFrame is locate in the same place as the breakpoint
         this.assertLocationMatchesExpected(topFrame, breakpoint);
 
-        if (stackTraceVerifier !== undefined) {
-            stackTraceVerifier.verify(stackTraceResponse);
+        if (typeof verifications.stackTrace === 'string') {
+            const assertions = new StackTraceStringAssertions(breakpoint);
+            assertions.assertResponseMatches(stackTraceResponse, verifications.stackTrace);
+        } else if (typeof verifications.stackTrace === 'object') {
+            const assertions = new StackTraceObjectAssertions(this._breakpointsWizard);
+            assertions.assertResponseMatches(stackTraceResponse, verifications.stackTrace);
         }
 
         if (verifications.variables !== undefined) {
@@ -124,25 +88,6 @@ export class BreakpointsAssertions {
         }
 
         await this._breakpointsWizard.resume();
-    }
-
-    public applyIgnores(formattedExpectedStackTrace: string, stackTrace: DebugProtocol.StackFrame[]): void {
-        const ignoreFunctionNameText = '<__IGNORE_FUNCTION_NAME__>';
-        const ignoreFunctionName = findLineNumber(formattedExpectedStackTrace, formattedExpectedStackTrace.indexOf(ignoreFunctionNameText));
-        if (ignoreFunctionName >= 0) {
-            expect(stackTrace.length).to.be.greaterThan(ignoreFunctionName);
-            const ignoredFrame = stackTrace[ignoreFunctionName];
-            ignoredFrame.name = `${ignoreFunctionNameText} [${ignoredFrame.source!.name}] Line ${ignoredFrame.line}`;
-        }
-    }
-
-    private extractStackTrace(stackTrace: DebugProtocol.StackFrame[]): string {
-        return stackTrace.map(f => this.printStackTraceFrame(f)).join('\n');
-    }
-
-    private printStackTraceFrame(frame: DebugProtocol.StackFrame): string {
-        let frameName = frame.name;
-        return `${frameName}:${frame.column}${frame.presentationHint && frame.presentationHint !== 'normal' ? ` (${frame.presentationHint})` : ''}`;
     }
 
     private assertLocationMatchesExpected(objectWithLocation: IObjectWithLocation, breakpoint: BreakpointWizard): void {
